Refact
-Refactoring of auxiliary check functions for user controllers. Division of files and folder for each.
