+++ src/controllers/user/userData.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
-import { findUserById, getUserData } from "../../utils/user/userValidations";
+import { findUserById } from "../../utils/user/search/findUserById";
+import { findUserData } from "../../utils/user/search/findUserData";
 
 export const userData = async (req: Request, res: Response) => {
   try {
@@ -9,7 +10,7 @@ export const userData = async (req: Request, res: Response) => {
       return res.status(404).send({ message: "No saves found!" });
     }
 
-    const userData = await getUserData(req.params.id);
+    const userData = await findUserData(req.params.id);
 
     return res.status(200).send(userData);
   } catch (error) {
