+++ src/utils/auth/emailRegex.ts
@@ -1,0 +1,1 @@
+export const emailRegex = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
