+++ src/utils/auth/hashPass.ts
@@ -1,0 +1,7 @@
+import bcrypt from "bcrypt";
+
+export const hashPass = async (pass: string) => {
+  const password = await bcrypt.hash(pass, 10);
+
+  return password;
+};
