+++ src/utils/user/search/findUserByEmail.ts
@@ -1,0 +1,9 @@
+import { userModel } from "../../../models/userModel";
+
+export const findUserByEmail = async (email: string) => {
+  const user = await userModel.findOne({ email });
+
+  if (!user) return false;
+
+  return user;
+};
